Replaces CreateInflowFileFromERARunoff.py CreateInflowFileFromERA5Runoff.py, which has the same functionality.
